requirements.txt
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
flask-restplus==0.12.1
Flask-Script==2.0.6
get==2018.11.19
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
MarkupSafe==1.1.1
post==2018.11.20
public==2018.11.20
pyrsistent==0.14.11
pytz==2018.9
PyYAML==5.1
query-string==2018.11.20
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
